+++ CHANGELOG.md
@@ -52,6 +52,7 @@ for soon-to-be removed features.
   - fix prefix suffix group style for native select
   - fix width of the native select when it is in a form-field (fill now 100% of the available place)
 - **grid**: an error was thrown in console when change current page (grid pagination)
+- **dropdown**: improve accessibility
 
 ### Security